+++ tabox/ta_func/ta_MIN.py
@@ -4,8 +4,9 @@ import numpy as np
 from .ta_utils import check_array, check_begidx1, check_timeperiod
 from ..retcode import TA_RetCode
 
+
 def min_double(left: cython.double, right: cython.double) -> cython.double:
-    """ min_double(left, right) -> double
+    """min_double(left, right) -> double
 
     Return the minimum of two doubles
     """
@@ -14,8 +15,9 @@ def min_double(left: cython.double, right: cython.double) -> cython.double:
     else:
         return right
 
-def TA_MIN_Lookback(optInTimePeriod: cython.int) -> cython.int:
-    """ TA_MIN_Lookback(optInTimePeriod) -> int
+
+def TA_MIN_Lookback(optInTimePeriod: cython.int) -> cython.Py_ssize_t:
+    """TA_MIN_Lookback(optInTimePeriod) -> int
 
     MIN Lookback
 
@@ -26,12 +28,19 @@ def TA_MIN_Lookback(optInTimePeriod: cython.int) -> cython.int:
     """
     return optInTimePeriod - 1
 
+
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def TA_MIN_LARGE_TIMEPERIOD(startIdx: cython.int, endIdx: cython.int, inReal: cython.double[::1], optInTimePeriod: cython.int, outReal: cython.double[::1]) -> cython.int:
+def TA_MIN_LARGE_TIMEPERIOD(
+    startIdx: cython.Py_ssize_t,
+    endIdx: cython.Py_ssize_t,
+    inReal: cython.double[::1],
+    optInTimePeriod: cython.int,
+    outReal: cython.double[::1],
+) -> cython.int:
     n: cython.int = endIdx - startIdx + 1
-    prefixMin: cython.double[::1] = np.empty(shape=(n,), dtype=np.float64) # [0] * n
+    prefixMin: cython.double[::1] = np.empty(shape=(n,), dtype=np.float64)  # [0] * n
     suffixMin: cython.double[::1] = np.empty(shape=(n,), dtype=np.float64)
 
     i: cython.int = 0
@@ -40,17 +49,17 @@ def TA_MIN_LARGE_TIMEPERIOD(startIdx: cython.int, endIdx: cython.int, inReal: cy
         if i % optInTimePeriod == 0:
             prefixMin[i] = inReal[inIdx]
         else:
-            prefixMin[i] = min_double(prefixMin[i-1], inReal[inIdx])
+            prefixMin[i] = min_double(prefixMin[i - 1], inReal[inIdx])
         i += 1
         inIdx += 1
 
-    i: cython.int = n-1
+    i: cython.int = n - 1
     inIdx: cython.int = startIdx + i
     while i >= 0:
         if i == n - 1 or (i + 1) % optInTimePeriod == 0:
             suffixMin[i] = inReal[inIdx]
         else:
-            suffixMin[i] = min_double(suffixMin[i+1], inReal[inIdx])
+            suffixMin[i] = min_double(suffixMin[i + 1], inReal[inIdx])
         i -= 1
         inIdx -= 1
 
@@ -63,11 +72,18 @@ def TA_MIN_LARGE_TIMEPERIOD(startIdx: cython.int, endIdx: cython.int, inReal: cy
 
     return TA_RetCode.TA_SUCCESS
 
+
 @cython.boundscheck(False)
 @cython.wraparound(False)
 @cython.cdivision(True)
-def TA_MIN_SMALL_TIMEPERIOD(startIdx: cython.int, endIdx: cython.int, inReal: cython.double[::1], optInTimePeriod: cython.int, outReal: cython.double[::1]) -> cython.int:
-    nbInitialElementNeeded: cython.int = optInTimePeriod-1
+def TA_MIN_SMALL_TIMEPERIOD(
+    startIdx: cython.Py_ssize_t,
+    endIdx: cython.Py_ssize_t,
+    inReal: cython.double[::1],
+    optInTimePeriod: cython.int,
+    outReal: cython.double[::1],
+) -> cython.int:
+    nbInitialElementNeeded: cython.int = optInTimePeriod - 1
 
     if startIdx < nbInitialElementNeeded:
         startIdx = nbInitialElementNeeded
@@ -99,19 +115,31 @@ def TA_MIN_SMALL_TIMEPERIOD(startIdx: cython.int, endIdx: cython.int, inReal: cy
             lowest = tmp
         outReal[outIdx] = lowest
         outIdx += 1
-        trailingIdx+=1
-        today+=1
-    
+        trailingIdx += 1
+        today += 1
+
     return TA_RetCode.TA_SUCCESS
 
-def TA_MIN(startIdx: cython.int, endIdx: cython.int, inReal: cython.double[::1], optInTimePeriod: cython.int, outReal: cython.double[::1]) -> cython.int:
+
+def TA_MIN(
+    startIdx: cython.Py_ssize_t,
+    endIdx: cython.Py_ssize_t,
+    inReal: cython.double[::1],
+    optInTimePeriod: cython.int,
+    outReal: cython.double[::1],
+) -> cython.int:
     if optInTimePeriod < 100:
-        return TA_MIN_SMALL_TIMEPERIOD(startIdx, endIdx, inReal, optInTimePeriod, outReal)
+        return TA_MIN_SMALL_TIMEPERIOD(
+            startIdx, endIdx, inReal, optInTimePeriod, outReal
+        )
     else:
-        return TA_MIN_LARGE_TIMEPERIOD(startIdx, endIdx, inReal, optInTimePeriod, outReal)
+        return TA_MIN_LARGE_TIMEPERIOD(
+            startIdx, endIdx, inReal, optInTimePeriod, outReal
+        )
+
 
 def MIN(real: np.ndarray, timeperiod: cython.int) -> np.ndarray:
-    """ MIN(real[, timeperiod=?])
+    """MIN(real[, timeperiod=?])
 
     Highest value over a specified period (Math Operators)
